feat: checkpoint 2
so got a lot of progress done, but im realizing i want to use an arena allocator
for the parser. otherwise the freeing fucking sucks. but i'll get there when i
get there. so far though, this implementation is much much easier and i feel
like it should support me better overall.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>