feat: pretty much there i think
so yeah we now properly parse out the directory structure, parse out all of the
files, then smush it into a json structure that even allows us to extend it
later with templates. very very good. i think we're pretty much at the point
where i can make a manager service and get the orchestrator to send the data up
to the manager. very nice

Signed-off-by: Matt Provost <mattprovost6@gmail.com>