feat: support environment variables
so now we support environment variables everywhere. this is super sick. super
important.

next up i think is gonna be custom functions from the wingman so it can
basically say "hey, i want to register these functions" and when a function call
is made, the request is sent to the wingman to evaluate. should work super well
and allow me to get all the dynamic functionality i need.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>