feat: use functions in the wingman
holy canoly. we got functions in the wingmen. now we can define a function to
get the token for the tunnel (ie create the tunnel if it doesnt exist, then
return the token for it, or create the tunnel when the wingman starts and just
return the token when it is called) and we can also register the function to get
a url and have the zone configured for the root of all of the domains in the
cluster. super sick.

Signed-off-by: Matt Provost <mattprovost6@gmail.com>